Edit question form bugfix 2
