+++ frontend/peerprep/app/@application/(main)/questions-management/add/page.tsx
@@ -1,10 +1,12 @@
 "use client";
 import AddQuestionForm from "@/components/addquestionform";
+import * as Y from "yjs";
 
 export default function AddQuestionsPage() {
+  const yDoc = new Y.Doc();
   return (
     <div className="flex">
-      <AddQuestionForm />
+      <AddQuestionForm yDoc={yDoc} />
     </div>
   );
 }
