+++ frontend/peerprep/app/@application/(main)/questions-management/edit/[id]/page.tsx
@@ -1,10 +1,12 @@
 "use client";
 import EditQuestionForm from "@/components/editquestionform";
+import * as Y from "yjs";
 
 export default function AddQuestionsPage() {
+  const yDoc = new Y.Doc();
   return (
     <div className="flex">
-      <EditQuestionForm />
+      <EditQuestionForm yDoc={yDoc} />
     </div>
   );
 }
