+++ frontend/peerprep/hooks/useQuestionForm.ts
@@ -1,4 +1,5 @@
-import { useState, useEffect } from 'react';
+import { editor } from 'monaco-editor';
+import { useState, useEffect, useRef } from 'react';
 import * as Y from "yjs";
 
 export function useQuestionForm(initialValues: {
@@ -17,23 +18,8 @@ export function useQuestionForm(initialValues: {
   const [templateCode, setTemplateCode] = useState(initialValues.templateCode || "");
   const [testCases, setTestCases] = useState(initialValues.testCases || [{ input: "", output: "" }]);
   const [language, setLanguage] = useState(initialValues.language || "javascript");
-
   // Yjs setup
-  const yDoc = new Y.Doc();
-  const yText = yDoc.getText("code");
-  const [YDocUpdate, setYDocUpdate] = useState<Uint8Array>(Y.encodeStateAsUpdateV2(yDoc));
 
-  const onMount = async (editor: any) => {
-    const model = editor.getModel();
-    if (model) {
-      const MonacoBinding = (await import("y-monaco")).MonacoBinding;
-      new MonacoBinding(yText, model, new Set([editor]));
-    }
-    model.setValue(templateCode);
-    yDoc.on("update", () => {
-      setYDocUpdate(Y.encodeStateAsUpdateV2(yDoc));
-    });
-  };
 
   return {
     title,
@@ -50,7 +36,5 @@ export function useQuestionForm(initialValues: {
     setTestCases,
     language,
     setLanguage,
-    YDocUpdate,
-    onMount,
   };
-} 
+}
